[RISCV] Emulate small vector loads/stores only 8 bytes at a time.
This patch removes the granular processing of the tail bytes of loads and
stores when we have VLEN=128, data size <= 64 bit, nf = 1 by 4, 2 and 1 byte
by doing one extra load/store of 8 bytes in one go.
This is because the tail bytes will be either ignored or set to 1 anyway and
we can so avoid extra if closes or loop iterations.

Signed-off-by: Paolo Savini <paolo.savini@embecosm.com>